SWC-6528: create user page fixtures, organize tests by subject rather than authentication state
